+++ src/segment_reshape/topology/find_related.py
@@ -17,14 +17,23 @@
 #  You should have received a copy of the GNU General Public License
 #  along with segment-reshape-qgis-plugin. If not, see <https://www.gnu.org/licenses/>.
 
-from typing import Iterable, Iterator, List, NamedTuple, Optional, Tuple, cast
+from typing import (
+    FrozenSet,
+    Iterable,
+    Iterator,
+    List,
+    NamedTuple,
+    Optional,
+    Set,
+    Tuple,
+    cast,
+)
 
 from qgis.core import (
     QgsFeature,
     QgsFeatureRequest,
     QgsGeometry,
     QgsLineString,
-    QgsMultiLineString,
     QgsPoint,
     QgsPointXY,
     QgsPolygon,
@@ -34,6 +43,9 @@ from qgis.core import (
 )
 
 from segment_reshape.geometry.reshape import ReshapeCommonPart, ReshapeEdge
+
+Point = Tuple[float, float]
+Segment = FrozenSet[Point]
 
 
 class CommonGeometriesResult(NamedTuple):
@@ -122,47 +134,38 @@ def _as_point_or_line_components(geom: QgsGeometry) -> Iterator[QgsGeometry]:
             yield QgsGeometry(part.clone())
 
 
-def _get_containing_component(
-    geom: QgsGeometry, contains_geom: QgsGeometry
-) -> QgsGeometry:
-    for component in _as_point_or_line_components(geom):
-        if component.contains(contains_geom):
-            return component
+def _get_geom_component_of_vertex(geom: QgsGeometry, vertex_id: int) -> QgsGeometry:
+    """Returns the geometry component that vertex_id belongs to.
 
-    raise ValueError(
-        f"could not find component of {geom} that contains {contains_geom}"
-    )
+    Geometry component is either a single point or linestring from multipoint or
+    multilinestring. In case of polygons geometry component is the edge linear ring
+    that vertex_id belongs to.
 
+    Args:
+        geom (QgsGeometry): Input geometry
+        vertex_id (int): Id of the vertex
 
-def _get_line_component_for_part_and_ring(
-    geom: QgsGeometry, part_index: int, ring_index: int
-) -> QgsGeometry:
-    try:
-        part = list(geom.constParts())[part_index]
-        if not isinstance(part, QgsPolygon):
-            return QgsGeometry(part.clone())
-        elif ring_index == 0:
-            return QgsGeometry(part.exteriorRing().clone())
-        else:
-            return QgsGeometry(part.interiorRing(ring_index - 1).clone())
-    except IndexError:
+    Raises:
+        ValueError: Raised if vertex_id is not found from the input geometry.
+
+    Returns:
+        QgsGeometry: The geometry component vertex_id belongs to
+    """
+
+    success, vertex_details = geom.vertexIdFromVertexNr(vertex_id)
+    if not success:
         raise ValueError(
-            f"could not extract part {part_index} ring {ring_index} of {geom}"
+            f"could not find vertex details by index {vertex_id}" f" from {geom}"
         )
 
-
-def _split_at_position(geom: QgsGeometry, position: QgsPoint) -> QgsGeometry:
-    new = QgsMultiLineString()
-    collected = QgsLineString()
-    for vertex in geom.vertices():
-        collected.addVertex(vertex.clone())
-        if vertex == position:
-            if collected.vertexCount() > 1:
-                new.addGeometry(collected)
-            collected = QgsLineString([vertex.clone()])
-    if collected.vertexCount() > 1:
-        new.addGeometry(collected)
-    return QgsGeometry(new)
+    part = list(geom.constParts())[vertex_details.part]
+    if isinstance(part, QgsPolygon):
+        if vertex_details.ring == 0:
+            return QgsGeometry(part.exteriorRing().clone())
+        else:
+            return QgsGeometry(part.interiorRing(vertex_details.ring - 1).clone())
+    else:
+        return QgsGeometry(part.clone())
 
 
 def _find_vertex_indices(geom: QgsGeometry, positions: Iterable[QgsPoint]) -> List[int]:
@@ -200,88 +203,67 @@ def get_common_geometries(
             f"unsupported source geometry type {main_feature.geometry().type()}"
         )
 
-    from_vertex, to_vertex = main_feature_segment
+    from_vertex_id, to_vertex_id = main_feature_segment
     trigger_geometry = main_feature.geometry()
-    trigger_segment = QgsGeometry.fromPolyline(
-        [
-            trigger_geometry.vertexAt(from_vertex).clone(),
-            trigger_geometry.vertexAt(to_vertex).clone(),
-        ]
-    )
-    success, to_vertex_id = trigger_geometry.vertexIdFromVertexNr(to_vertex)
-    if not success:
-        raise ValueError(
-            f"could not find vertex details by index {to_vertex}"
-            f" from {trigger_geometry}"
-        )
 
-    # start with the full line component as the result
-    result = _get_line_component_for_part_and_ring(
-        trigger_geometry, to_vertex_id.part, to_vertex_id.ring
+    from_vertex = trigger_geometry.vertexAt(from_vertex_id).clone()
+    to_vertex = trigger_geometry.vertexAt(to_vertex_id).clone()
+
+    # edge that was clicked
+    trigger_segment = frozenset(
+        (
+            (from_vertex.x(), from_vertex.y()),
+            (to_vertex.x(), to_vertex.y()),
+        )
     )
+
+    # line component that was clicked
+    trigger_part = _get_geom_component_of_vertex(trigger_geometry, to_vertex_id)
+    # start with the full line component as the result
+    line_segments_to_keep = _as_line_segments(trigger_part)
 
     common_part_candidates: List[Tuple[QgsVectorLayer, QgsFeature]] = []
     possible_edge_candidates: List[Tuple[QgsVectorLayer, QgsFeature, QgsGeometry]] = []
 
     for layer, feature in related_features_by_layer:
         for component in _as_point_or_line_components(feature.geometry()):
-
-            # found a point which can only act as a break to the segment,
-            # split the result to multipart line at the point if possible
+            # found a point which can only act as a break to the segment
             # -> collect as possible edge
             if component.type() == QgsWkbTypes.GeometryType.PointGeometry:
-                result = _split_at_position(result, component.vertexAt(0))
                 possible_edge_candidates.append((layer, feature, component))
+                continue
 
-            # found a line that shares the trigger segment, intersection to
-            # reduce the result to the part that is shared as common segment
+            component_segments = _as_line_segments(component)
+
+            # Found a line that contains the trigger segment.
+            # Keep intersection to reduce the result to the part that is shared
+            # as common segment.
             # -> collect as common part
-            elif component.contains(trigger_segment):
-                # TODO: how to handle partial linear intersection where the
-                # component contains the trigger, but the containment is either
-                # within one longer segment or multiple shorter segments, and
-                # vertices will not match (same pair in same or reversed order)
-                result = result.intersection(component)
-                # intersection along multiple segments will return each segment
-                # as a separate part, merge parts here to keep the continuous line
-                if result.isMultipart():
-                    result = result.mergeLines()
+            if trigger_segment in component_segments:
+                line_segments_to_keep = line_segments_to_keep.intersection(
+                    component_segments
+                )
                 common_part_candidates.append((layer, feature))
 
-            # found a line that may intersect along the result but not along
-            # trigger segment, cross or only touch the result, difference to reduce
-            # the result by removing parts of intersection along the result or
-            # by breaking the result to parts at the crossings or touch locations
+            # Found a line that does not contain trigger segment so remove it from
+            # the result.
+            # It might still touch the result
             # -> collect as possible edge
             else:
-                # TODO: how to handle partial linear intersection where the result
-                # is reduced possibly even along the trigger segment, if the component
-                # has an vertex that falls along a segment of the result
-                # TODO: how to handle non-vertex touch/cross where the result will
-                # be split on the touching vertex or crossing segment even if there
-                # is no vertex on result
-                result = result.difference(component)
+                line_segments_to_keep = line_segments_to_keep.difference(
+                    component_segments
+                )
                 possible_edge_candidates.append((layer, feature, component))
 
-            # TODO: how to handle ring-like lines which may result in a split
-            # at a valid split point, and an existing split at the ring boundary,
-            # and after the split or difference operation we want to keep the
-            # continuous line around the boundary with with split or parts removed,
-            # since this conflicts with the points or difference operation results
-            # for crossing on touching lines which only result in breaks and will
-            # be rejoined if mergeLines is ran without special checks
-            # if result.isMultipart():
-            #     result = result.mergeLines()
+    segment = _build_line_from_line_segment_set(
+        trigger_part,
+        line_segments_to_keep,
+        trigger_segment,
+        edge_candidate_geometries=[
+            geometry for _, _, geometry in possible_edge_candidates
+        ],
+    )
 
-            # if result is now multipart due to the geometry operations,
-            # only keep the component which contains the trigger segment
-            # for polygon rings the two parts might overlap with
-            # the loop boundary, resulting in two continuous parts
-            if result.isMultipart():
-                result = _get_containing_component(result, trigger_segment)
-
-    # result is now the longest common combined segment available from the trigger
-    segment = cast(QgsLineString, result.constGet().clone())
     start, end = QgsGeometry(segment.startPoint()), QgsGeometry(segment.endPoint())
 
     common_parts: List[ReshapeCommonPart] = [
@@ -329,3 +311,82 @@ def get_common_geometries(
             )
 
     return CommonGeometriesResult(segment, common_parts, edges)
+
+
+def _as_line_segments(geometry: QgsGeometry) -> Set[Segment]:
+    vertices = [
+        (point.x(), point.y()) for point in cast(QgsLineString, geometry.asPolyline())
+    ]
+    return {frozenset(line) for line in zip(vertices, vertices[1:])}
+
+
+def _build_line_from_line_segment_set(
+    trigger_part: QgsGeometry,
+    line_segments_to_keep: Set[Segment],
+    trigger_segment: Segment,
+    edge_candidate_geometries: List[QgsGeometry],
+) -> QgsLineString:
+    """Build a subline from trigger_part with constraints
+
+    Builds a line which contains the trigger_segment and has only segments from
+    line_segments_to_keep set and which is split at points where any geometry
+    from edge_candidate_geometries touches the line.
+
+    Args:
+        trigger_part (QgsGeometry): Input geometry from where the subline is extracted
+        line_segments_to_keep (Set[Segment]): Set of segments that the subline is build
+            from
+        trigger_segment (Segment): Edge which was clicked
+        edge_candidate_geometries (List[QgsGeometry]): List of geometries which might
+            act as break points
+
+    Returns:
+        QgsLineString: A sublinestring from the input geometry that follows the
+            constraints
+    """
+
+    possible_split_points = {
+        (vertex.x(), vertex.y())
+        for geom in edge_candidate_geometries
+        for vertex in geom.vertices()
+    } & {(vertex.x(), vertex.y()) for vertex in trigger_part.vertices()}
+
+    vertex_iterator = trigger_part.vertices()
+    next_vertex_iterator = trigger_part.vertices()
+    _ = next(next_vertex_iterator)  # advance the next_vertex iterator by one
+
+    parts: List[List[QgsPoint]] = []
+    current_part_vertices: List[QgsPoint] = []
+    for vertex, next_vertex in zip(vertex_iterator, next_vertex_iterator):
+        vertex_tuple = (vertex.x(), vertex.y())
+        next_vertex_tuple = (next_vertex.x(), next_vertex.y())
+        segment = frozenset((vertex_tuple, next_vertex_tuple))
+
+        if segment in line_segments_to_keep:
+            if segment == trigger_segment:
+                trigger_in_part = len(parts)  # this will be the index of this part
+
+            if not current_part_vertices:
+                current_part_vertices.append(QgsPoint(*vertex_tuple))
+            current_part_vertices.append(QgsPoint(*next_vertex_tuple))
+        elif current_part_vertices:
+            parts.append(current_part_vertices)
+            current_part_vertices = []
+
+        if next_vertex_tuple in possible_split_points and current_part_vertices:
+            parts.append(current_part_vertices)
+            current_part_vertices = []
+
+    if current_part_vertices:
+        parts.append(current_part_vertices)
+
+    if _is_linear_ring_split(parts) and trigger_in_part in (0, len(parts) - 1):
+        # The original geometry was a linear ring and was split at middle.
+        # The trigger segment is in the first or last part so those must be merged.
+        return QgsLineString(parts[-1] + parts[0][1:])
+    else:
+        return QgsLineString(parts[trigger_in_part])
+
+
+def _is_linear_ring_split(parts: List[List[QgsPoint]]) -> bool:
+    return len(parts) >= 2 and parts[0][0] == parts[-1][-1]